+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+- Evaluate FMS through reusable evaluation pipeline
 - Don't take ownership of wrapped types passed from JavaScript
 
 ### Removed